rewrite so that methods are defined in a trait
This will make it much easier to handle inheriting methods from
superclasses. In particular, the macro does not have to know the
names/types of the methods of superclasses, which would let them
compose much more easily.

Unfortunately, it means that the trait *and* struct must be imported
-- as must the trait and struct from superclasses! This may be a
non-starter. If only we had inherent traits, though, we could overcome
this limitation.