allow(deprecated) for glib::CallbackGuard, until we formally remove it
Rust 1.24 made it so that a panic!() won't unwind across FFI
boundaries: we won't unwind if we panic inside a Rust function
declared extern "C".

https://github.com/rust-lang/rust/pull/46833

However, gnome-class does not formally mandate a particular Rust
version; we've been running on the assumption that we are running on
nightly, or "recent enough".  For now, just supress the deprecation
warnings from glib-rs, until we formalize our requirements for the
rustc version.